dock panel in own module
